+++ src/services/order/addOrder.ts
@@ -1,0 +1,5 @@
+const addOrder = async () => {
+  return 'addOrder';
+};
+
+export default addOrder;
